+++ zetaclient/chains/sui/observer/outbound.go
@@ -45,15 +45,12 @@ func (ob *Observer) ProcessOutboundTrackers(ctx context.Context) error {
 			continue
 		}
 
-		logger := ob.Logger().Outbound.With().
-			Str(logs.FieldMethod, "ProcessOutboundTrackers").
-			Uint64(logs.FieldNonce, nonce).
-			Logger()
+		logger := ob.Logger().Outbound.With().Uint64(logs.FieldNonce, nonce).Logger()
 
 		// should not happen
 		if len(tracker.HashList) == 0 {
 			// we don't want to block other cctxs, so let's error and continue
-			logger.Error().Str(logs.FieldTracker, tracker.Index).Msg("tracker hash list is empty")
+			logger.Error().Str("tracker_id", tracker.Index).Msg("tracker hash list is empty")
 			continue
 		}
 
@@ -92,8 +89,11 @@ func (ob *Observer) VoteOutbound(ctx context.Context, cctx *cctypes.CrossChainTx
 
 	// used checkpoint instead of block height
 	checkpoint, err := strconv.ParseUint(tx.Checkpoint, 10, 64)
-	if err != nil || checkpoint == 0 {
+	if err != nil {
 		return errors.Wrapf(err, "invalid checkpoint: %s", tx.Checkpoint)
+	}
+	if checkpoint == 0 {
+		return errors.New("checkpoint is zero")
 	}
 
 	// parse outbound event
@@ -234,8 +234,9 @@ func (ob *Observer) postVoteOutbound(ctx context.Context, msg *cctypes.MsgVoteOu
 	case zetaTxHash != "":
 		ob.Logger().Outbound.Info().
 			Str(logs.FieldTx, msg.ObservedOutboundHash).
+			Uint64(logs.FieldNonce, msg.OutboundTssNonce).
 			Str(logs.FieldZetaTx, zetaTxHash).
-			Str(logs.FieldBallot, ballot).
+			Str(logs.FieldBallotIndex, ballot).
 			Msg("posted outbound vote")
 	}